account_service/requirements.txt
Flask==3.0.0
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
mysql-connector-python==8.3.0